refactor(routes): decompose calculator and plot page god components
Extract long-lived `$effect` orchestrations from the two largest files in
the repo into headless setup functions under `$lib/state/`, mirroring the
pattern established by PR #602 (`setupCalculatorUrlSync`), #603 (`app-init`),
and #604 (result-table decomposition).

- `setupMultiProgramCalculation` — built-in + external multi-program calc
- `setupMultiEntityCalculation` — multi-material / multi-particle calc
- `setupInverseRangeCalculation` / `setupInverseStpCalculation`
- `setupPlotPreviewCalculation` — auto-preview series for the plot page
- `setupPlotUrlRestore` — URL → state hydration including series restore

Pure helpers moved out of the page scripts:

- `$lib/utils/inverse-units.ts` (range / STP unit conversions)
- `$lib/utils/energy-anchor-options.ts` (unit anchor strip option builder)
- `$lib/utils/calculator-pdf-metadata.ts` (advanced PDF metadata)
- `$lib/export/plot-image.ts` (SVG / PNG download with JSROOT + DOM fallbacks)

Also dedupe `resolveExtLocalIdForLabel` (three near-identical copies) into
`$lib/external-data/ids.ts`.

Page line counts:

- `routes/calculator/+page.svelte`: 1815 → 1055 (-42%)
- `routes/plot/+page.svelte`: 1280 → 802 (-37%)

Behaviour is unchanged: every snapshot pattern, race-protection flag, and
reactivity workaround was preserved verbatim during extraction. `pnpm run
check`, `lint`, and `format:check` all clean. 1471/1474 unit tests pass
(the 3 pre-existing `guard-forbidden-files.test.ts` failures are git
signing infra failures noted in earlier session logs, unrelated to this
change). Manual SSR smoke for `/calculator` and `/plot` returns clean
HTML.
