+++ docs/ai-logs/README.md
@@ -27,6 +27,7 @@ summary table of all sessions.
 
 | File                                                                                                                         | Date         | Topic                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         |
 | ---------------------------------------------------------------------------------------------------------------------------- | ------------ | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| [2026-05-27-page-decomposition.md](2026-05-27-page-decomposition.md)                                                          | 27 May 2026  | **Decompose `/calculator` and `/plot` page components** (Claude Sonnet 4.6 via Claude Code): Extracted four big `$effect` blocks from `routes/calculator/+page.svelte` (multi-program calc, multi-entity calc, inverse-range calc, inverse-stp calc) and two from `routes/plot/+page.svelte` (preview calc, URL restore) into headless `setup*()` state modules under `$lib/state/`. Extracted pure helpers (`inverse-units`, `energy-anchor-options`, `calculator-pdf-metadata`, `plot-image`) and deduplicated `resolveExtLocalIdForLabel` into `$lib/external-data/ids.ts`. Calculator page 1815 → 1055 lines (-42%); plot page 1280 → 802 lines (-37%). Static checks + lint + format all clean. |
 | [2026-05-26-issue-615-particle-tab-decompose.md](2026-05-26-issue-615-particle-tab-decompose.md)                            | 26 May 2026  | **Decompose particle-tab.svelte (#615)** (Claude Opus 4.7 via Claude Code): split the monolithic particle picker into a thin `particle-tab.svelte` orchestrator plus `particle-list-view.svelte`, `particle-grid-view.svelte`, and a shared `particle-tab-helpers.ts` module; preserved every `data-testid`, `aria-selected`, `aria-pressed`, and `data-available` contract so existing Playwright E2E selectors keep working. |
 | [2026-05-26-pr614-review-thread-fixes.md](2026-05-26-pr614-review-thread-fixes.md)                                          | 26 May 2026  | **PR #614 review-thread fixes + E2E follow-up** (GPT-5.5 via GitHub Copilot coding agent): aligned the particle picker with the responsive scan-grid contract (including 44px mobile tap targets), fixed particle-grid empty-state behavior for external-only matches, canonicalized calculator URL `across` to plural tokens with singular decode aliases, gated `across` restore on valid matching lists, and tightened `calculator-url.spec.ts` / `particle-grid-view.spec.ts` assertions for deterministic CI coverage. |
 | [2026-05-26-issue-599-periodic-grid-particle-picker.md](2026-05-26-issue-599-periodic-grid-particle-picker.md)              | 26 May 2026  | **Issue #599 — Periodic-grid scan view in advanced particle picker** (Claude Opus 4.7 via Claude Code): Added a list/grid view toggle gated on Advanced mode in `particle-tab.svelte`. Grid renders Z=1..118 as Z+symbol tiles in a `grid-cols-6 sm:grid-cols-9` responsive layout, with electron omitted, unavailable tiles at opacity 0.4, and external-only particles in a wrap row below the grid. Tap targets meet 44px mobile a11y minimum. New `tests/e2e/particle-grid-view.spec.ts` covers desktop and mobile flows. |
